backend/pinecone_integration/requirements.txt
pinecone-client
pypdf2
python-dotenv
